# Improve NEON64 decoding speed