@aklomp 2016-10-20
This is really clever, good work! I'll review it further tomorrow or in the weekend.