@aklomp 2016-10-22
Thanks, merged. I rebased the commit onto Master so that I could do a fast-forward merge.